Create CP_003.cpp
WAP to find the largest rectangular area possible in a given histogram where the largest rectangle can be made of a number of contiguous bars. 

Example: 
For n = 5, [1, 2, 3, 4, 5]
The largest rectangle can be obtained of breadth=3 and length=3. Its starting index is 2 and ending index is 4 and it has a height of 3.
Hence, area = 3x3 = 9

Reference: https://leetcode.com/problems/largest-rectangle-in-histogram/